+++ src/components/animate-split/index.ts
@@ -1,0 +1,2 @@
+export { MPAnimateSplit } from './MPAnimateSplit';
+export type { MPAnimateSplitBy, MPAnimateSplitProps } from './MPAnimateSplit';